pyproject.toml
[project]
name = "agentic-data-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.129.0",
    "google-genai>=1.64.0",
    "httpx>=0.28.1",
    "kaleido>=1.2.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "plotly>=6.5.2",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "python-pptx>=1.0.2",
    "python-ulid>=3.1.0",
    "redis>=7.2.0",
    "reportlab>=4.4.10",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
]
